+++ packages/adapters/system/web-crypto-uuid/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @multitenantkit/adapter-system-web-crypto
 
+## 0.2.6
+
+### Patch Changes
+
+- Add debug logging for customFields propagation
+
 ## 0.2.5
 
 ### Patch Changes